Fix errors after renaming
